Admin hide/show handlers renaming
