+++ EOLib/PacketHandlers/AdminInteract/AdminInteractAgree.cs
@@ -5,20 +5,23 @@ using EOLib.Domain.Map;
 using EOLib.Net;
 using EOLib.Net.Handlers;
 
-namespace EOLib.PacketHandlers
+namespace EOLib.PacketHandlers.AdminInteract
 {
+    /// <summary>
+    /// Admin unhiding
+    /// </summary>
     [AutoMappedType]
-    public class AdminHideHandler : InGameOnlyPacketHandler
+    public class AdminInteractAgree : InGameOnlyPacketHandler
     {
         private readonly ICharacterRepository _characterRepository;
         private readonly ICurrentMapStateRepository _currentMapStateRepository;
 
         public override PacketFamily Family => PacketFamily.AdminInteract;
-        public override PacketAction Action => PacketAction.Remove;
+        public override PacketAction Action => PacketAction.Agree;
 
-        public AdminHideHandler(IPlayerInfoProvider playerInfoProvider,
-                                ICharacterRepository characterRepository,
-                                ICurrentMapStateRepository currentMapStateRepository)
+        public AdminInteractAgree(IPlayerInfoProvider playerInfoProvider,
+                                  ICharacterRepository characterRepository,
+                                  ICurrentMapStateRepository currentMapStateRepository)
             : base(playerInfoProvider)
         {
             _characterRepository = characterRepository;
@@ -30,14 +33,14 @@ namespace EOLib.PacketHandlers
             var id = packet.ReadShort();
 
             if (id == _characterRepository.MainCharacter.ID)
-                _characterRepository.MainCharacter = Hidden(_characterRepository.MainCharacter);
+                _characterRepository.MainCharacter = Shown(_characterRepository.MainCharacter);
             else
             {
                 if (_currentMapStateRepository.Characters.ContainsKey(id))
                 {
                     var character = _currentMapStateRepository.Characters[id];
 
-                    var updatedCharacter = Hidden(character);
+                    var updatedCharacter = Shown(character);
                     _currentMapStateRepository.Characters[id] = updatedCharacter;
                 }
                 else
@@ -49,9 +52,9 @@ namespace EOLib.PacketHandlers
             return true;
         }
 
-        private static Character Hidden(Character input)
+        private static Character Shown(Character input)
         {
-            var renderProps = input.RenderProperties.WithIsHidden(true);
+            var renderProps = input.RenderProperties.WithIsHidden(false);
             return input.WithRenderProperties(renderProps);
         }
     }
